requirements.txt
scipy==1.6.3
Flask==2.0.1
opencv_python==4.5.2.54
requests==2.25.1
pandas==1.2.4
typing_extensions==3.7.4.3
Flask_Caching==1.9.0
eagerpy==0.30.0
matplotlib==3.3.4
tqdm==4.61.1
scikit_image==0.18.1
torch==1.7.1
torchvision==0.8.2
Werkzeug==2.0.1
Jinja2==3.0.0
redis==3.5.3
ipython==7.28.0
numba==0.54.0
Pillow~=8.2.0
pyDOE2==1.3.0
scikit_learn==1.0
tensorboardX==2.4
numpy==1.20.2
scikit-image~=0.18.1
